@@ -2,12 +2,12 @@ name: flutter_chat_types
 description: >
   Utility library for the flutter_chat_ui and flutter_firebase_chat_core libraries
   which contains shared type declarations.
-version: 3.5.0
+version: 3.6.0
 homepage: https://flyer.chat
 repository: https://github.com/flyerhq/flutter_chat_types
 
 environment:
-  sdk: '>=2.17.0 <3.0.0'
+  sdk: '>=2.18.0 <3.0.0'
 
 dependencies:
   equatable: ^2.0.5
@@ -15,7 +15,7 @@ dependencies:
   meta: ^1.8.0
 
 dev_dependencies:
-  build_runner: ^2.2.1
-  dart_code_metrics: ^4.19.1
+  build_runner: ^2.3.2
+  dart_code_metrics: ^4.21.3
   flutter_lints: ^2.0.1
-  json_serializable: ^6.4.1
+  json_serializable: ^6.5.4
