@@ -1,30 +1,25 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-part of 'partial_audio.dart';
+part of 'partial_video.dart';
 
 // **************************************************************************
 // JsonSerializableGenerator
 // **************************************************************************
 
-PartialAudio _$PartialAudioFromJson(Map<String, dynamic> json) => PartialAudio(
-      duration: Duration(microseconds: json['duration'] as int),
+PartialVideo _$PartialVideoFromJson(Map<String, dynamic> json) => PartialVideo(
+      height: (json['height'] as num?)?.toDouble(),
       metadata: json['metadata'] as Map<String, dynamic>?,
-      mimeType: json['mimeType'] as String?,
       name: json['name'] as String,
       repliedMessage: json['repliedMessage'] == null
           ? null
           : Message.fromJson(json['repliedMessage'] as Map<String, dynamic>),
       size: json['size'] as num,
       uri: json['uri'] as String,
-      waveForm: (json['waveForm'] as List<dynamic>?)
-          ?.map((e) => (e as num).toDouble())
-          .toList(),
+      width: (json['width'] as num?)?.toDouble(),
     );
 
-Map<String, dynamic> _$PartialAudioToJson(PartialAudio instance) {
-  final val = <String, dynamic>{
-    'duration': instance.duration.inMicroseconds,
-  };
+Map<String, dynamic> _$PartialVideoToJson(PartialVideo instance) {
+  final val = <String, dynamic>{};
 
   void writeNotNull(String key, dynamic value) {
     if (value != null) {
@@ -32,12 +27,12 @@ Map<String, dynamic> _$PartialAudioToJson(PartialAudio instance) {
     }
   }
 
+  writeNotNull('height', instance.height);
   writeNotNull('metadata', instance.metadata);
-  writeNotNull('mimeType', instance.mimeType);
   val['name'] = instance.name;
   writeNotNull('repliedMessage', instance.repliedMessage?.toJson());
   val['size'] = instance.size;
   val['uri'] = instance.uri;
-  writeNotNull('waveForm', instance.waveForm);
+  writeNotNull('width', instance.width);
   return val;
 }
